// WYNKWeb/src/app/Models/RegistrationMasterWebModel.ts
export class Registration_Master {
  UIN: string;
  CMPID: number;
  DateofRegistration: Date;
  Name: string;
  DateofBirth: Date;
  Gender: string;
  Address1: string;
  Address2: string;
  Address3: string;
  LocationID: string;
  FatherHusbandName: string;
  Occupation: string;
  Phone: string;
  AadharNumber: string;
  Age: number;


  PanCardNo: string;
  DrivingLicenseNo: string;
  PassportNo: string;
  IsForeignNational: number;

  EmailID: string;
  SourceofReferralID: number;
  ReferralName: string;
  PhotoPath: string;
  DoctorclinicName: string;
  BiometricPath: string;
  UpdatedBy: number;
  CreatedBy: number;
  IsDeleted: Boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date;
  MiddleName: string;
  LastName: string;
  AlternatePhoneNumber: string;
  AlternateMailID: string;
  PreferredLanguage: string;
  MaritalStatus: number;
  Insurance: Boolean;
  RegType: string;
}