// WYNKWeb/src/app/Models/RegistrationTranMasterWebModel .ts
export class RegistrationTran_Master {

  RegistrationTranID: number;
  UIN: string;
  DateofVisit: Date;
  TypeofVisit: number ;
  Status: number;
  StatusDateTime: Date ;
  Remarks: string;
  FollowupDate: Date ;
  IsDeleted: boolean;
  CreatedUTC: Date;
  UpdatedUTC: Date ;
  CreatedBy: number;
  UpdatedBy: number ;
  DoctorID: number ;
  Assigneddate: Date;
  //ConsulationFees: number;
  RegistrationFees: number;

}