// WYNKWeb/src/app/Models/PaymentWebModel .ts
export class Payment_Master {

  UIN: string;
  PaymentFor: string;
  PaymentAgainstID: number;
  PaymentReferenceID: number;
  PaymentMode: string;
  InstrumentNumber: string;
  Instrumentdate: Date;
  BankName: string;
  BankBranch: string;
  Expirydate: Date;
  Amount: number;
  InVoiceNumber: string;
  InVoiceDate: Date;
  AdvAmount: number;
  CreatedBy: number;
  UpdatedBy: string;
  UpdatedUTC: Date;
  TransactionID: number;
  ReceiptDate: Date;

}